4.0.3 fix warning

+++ Tests/iOS Tests/Helpers/NSString+PNTest.m
@@ -27,8 +27,8 @@
     NSMutableData *data= [NSMutableData new];
     unsigned char whole_byte;
     char byte_chars[3] = {'\0','\0','\0'};
-    int i = 0;
-    int length = string.length;
+    NSUInteger i = 0;
+    NSUInteger length = string.length;
     while (i < length-1) {
         char c = [string characterAtIndex:i++];
         if (c < '0' || (c > '9' && c < 'a') || c > 'f')
